class Assert changed to namespace.
Assert2: added doubleEqual.

Assert2: added complexEqual.

Assert2: added intEqual.

Assert2: added stringEqual.

Assert2: added true_.

Asser2: added false_.

Assert2: private functions moved out.

Assert2 -> Assert.

Assert under radsim.
